Buildmode

+++ code/modules/admin/verbs/buildmode.dm
@@ -1,57 +1,66 @@
-/datum/admins/proc/build_mode()
-	set name = "Toggle Build Mode"
+/client/proc/build_mode()
+	set name = "Toggle Build Mode Self"
 	set category = "Fun"
 
 	if(!check_rights(R_FUN))
 		return
 
-
-	if(usr.client.buildmode)
-		log_admin("[key_name(usr)] has left build mode.")
-		usr.client.buildmode = FALSE
-		usr.client.show_popup_menus = TRUE
-		for(var/obj/effect/bmode/buildholder/H)
-			if(H.cl == usr.client)
-				del(H)
-	else
-		log_admin("[key_name(usr)] has entered build mode.")
-		usr.client.buildmode = TRUE
-		usr.client.show_popup_menus = FALSE
-
-		var/obj/effect/bmode/buildholder/H = new/obj/effect/bmode/buildholder()
-		var/obj/effect/bmode/builddir/A = new/obj/effect/bmode/builddir(H)
-		A.master = H
-		var/obj/effect/bmode/buildhelp/B = new/obj/effect/bmode/buildhelp(H)
-		B.master = H
-		var/obj/effect/bmode/buildmode/C = new/obj/effect/bmode/buildmode(H)
-		C.master = H
-		var/obj/effect/bmode/buildquit/D = new/obj/effect/bmode/buildquit(H)
-		D.master = H
-
-		H.builddir = A
-		H.buildhelp = B
-		H.buildmode = C
-		H.buildquit = D
-		usr.client.screen += A
-		usr.client.screen += B
-		usr.client.screen += C
-		usr.client.screen += D
-		H.cl = usr.client
+	togglebuildmode(usr)
 
 
-/obj/effect/bmode
-	density = TRUE
-	anchored = TRUE
+/proc/togglebuildmode(mob/M as mob in player_list)
+	set name = "Toggle Build Mode"
+	set category = "Special Verbs"
+
+	if(!check_rights(R_FUN))
+		return
+
+	if(M.client)
+		if(M.client.buildmode)
+			log_admin("[key_name(usr)] has left build mode.")
+			message_admins("[ADMIN_TPMONTY(usr)] has left build mode.")
+			M.client.buildmode = FALSE
+			M.client.show_popup_menus = TRUE
+			for(var/obj/effect/bmode/buildholder/H)
+				if(H.cl == M.client)
+					del(H)
+		else
+			log_admin("[key_name(usr)] has entered build mode.")
+			message_admins("[ADMIN_TPMONTY(usr)] has entered build mode.")
+			M.client.buildmode = TRUE
+			M.client.show_popup_menus = FALSE
+
+			var/obj/effect/bmode/buildholder/H = new/obj/effect/bmode/buildholder()
+			var/obj/effect/bmode/builddir/A = new/obj/effect/bmode/builddir(H)
+			A.master = H
+			var/obj/effect/bmode/buildhelp/B = new/obj/effect/bmode/buildhelp(H)
+			B.master = H
+			var/obj/effect/bmode/buildmode/C = new/obj/effect/bmode/buildmode(H)
+			C.master = H
+			var/obj/effect/bmode/buildquit/D = new/obj/effect/bmode/buildquit(H)
+			D.master = H
+
+			H.builddir = A
+			H.buildhelp = B
+			H.buildmode = C
+			H.buildquit = D
+			M.client.screen += A
+			M.client.screen += B
+			M.client.screen += C
+			M.client.screen += D
+			H.cl = M.client
+
+/obj/effect/bmode//Cleaning up the tree a bit
+	density = 1
+	anchored = 1
 	layer = ABOVE_HUD_LAYER
 	dir = NORTH
 	icon = 'icons/misc/buildmode.dmi'
 	var/obj/effect/bmode/buildholder/master = null
 
-
 /obj/effect/bmode/builddir
 	icon_state = "build"
 	screen_loc = "NORTH,WEST"
-
 
 /obj/effect/bmode/builddir/clicked()
 	switch(dir)
@@ -65,64 +74,59 @@
 			dir = NORTHWEST
 		if(NORTHWEST)
 			dir = NORTH
-	return TRUE
-
+	return 1
 
 /obj/effect/bmode/buildhelp
 	icon = 'icons/misc/buildmode.dmi'
 	icon_state = "buildhelp"
 	screen_loc = "NORTH,WEST+1"
 
-
 /obj/effect/bmode/buildhelp/clicked()
 	switch(master.cl.buildmode)
 		if(1)
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
-			to_chat(usr, "<span class='notice'> Left Mouse Button        = Construct / Upgrade</span>")
-			to_chat(usr, "<span class='notice'> Right Mouse Button       = Deconstruct / Delete / Downgrade</span>")
-			to_chat(usr, "<span class='notice'> Left Mouse Button + ctrl = R-Window</span>")
-			to_chat(usr, "<span class='notice'> Left Mouse Button + alt  = Airlock</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>Left Mouse Button        = Construct / Upgrade</span>")
+			to_chat(usr, "<span class='boldnotice'>Right Mouse Button       = Deconstruct / Delete / Downgrade</span>")
+			to_chat(usr, "<span class='boldnotice'>Left Mouse Button + ctrl = R-Window</span>")
+			to_chat(usr, "<span class='boldnotice'>Left Mouse Button + alt  = Airlock</span>")
 			to_chat(usr, "")
-			to_chat(usr, "<span class='notice'> Use the button in the upper left corner to</span>")
-			to_chat(usr, "<span class='notice'> change the direction of built objects.</span>")
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>Use the button in the upper left corner to</span>")
+			to_chat(usr, "<span class='boldnotice'>change the direction of built objects.</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
 		if(2)
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
-			to_chat(usr, "<span class='notice'> Right Mouse Button on buildmode button = Set object type</span>")
-			to_chat(usr, "<span class='notice'> Middle Mouse Button on turf/obj        = Set object type</span>")
-			to_chat(usr, "<span class='notice'> Left Mouse Button on turf/obj          = Place objects</span>")
-			to_chat(usr, "<span class='notice'> Right Mouse Button                     = Delete objects</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>Right Mouse Button on buildmode button = Set object type</span>")
+			to_chat(usr, "<span class='boldnotice'>Middle Mouse Button on turf/obj        = Set object type</span>")
+			to_chat(usr, "<span class='boldnotice'>Left Mouse Button on turf/obj          = Place objects</span>")
+			to_chat(usr, "<span class='boldnotice'>Right Mouse Button                     = Delete objects</span>")
 			to_chat(usr, "")
-			to_chat(usr, "<span class='notice'> Use the button in the upper left corner to</span>")
-			to_chat(usr, "<span class='notice'> change the direction of built objects.</span>")
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>Use the button in the upper left corner to</span>")
+			to_chat(usr, "<span class='boldnotice'>change the direction of built objects.</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
 		if(3)
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
-			to_chat(usr, "<span class='notice'> Right Mouse Button on buildmode button = Select var(type) & value</span>")
-			to_chat(usr, "<span class='notice'> Left Mouse Button on turf/obj/mob      = Set var(type) & value</span>")
-			to_chat(usr, "<span class='notice'> Right Mouse Button on turf/obj/mob     = Reset var's value</span>")
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>Right Mouse Button on buildmode button = Select var(type) & value</span>")
+			to_chat(usr, "<span class='boldnotice'>Left Mouse Button on turf/obj/mob      = Set var(type) & value</span>")
+			to_chat(usr, "<span class='boldnotice'>Right Mouse Button on turf/obj/mob     = Reset var's value</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
 		if(4)
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
-			to_chat(usr, "<span class='notice'> Left Mouse Button on turf/obj/mob      = Select</span>")
-			to_chat(usr, "<span class='notice'> Right Mouse Button on turf/obj/mob     = Throw</span>")
-			to_chat(usr, "<span class='notice'> ***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
+			to_chat(usr, "<span class='boldnotice'>Left Mouse Button on turf/obj/mob      = Select")
+			to_chat(usr, "<span class='boldnotice'>Right Mouse Button on turf/obj/mob     = Throw")
+			to_chat(usr, "<span class='boldnotice'>***********************************************************</span>")
 	return TRUE
-
 
 /obj/effect/bmode/buildquit
 	icon_state = "buildquit"
 	screen_loc = "NORTH,WEST+3"
 
-
 /obj/effect/bmode/buildquit/clicked()
-	//togglebuildmode(master.cl.mob)
+	togglebuildmode(master.cl.mob)
 	return TRUE
 
-
 /obj/effect/bmode/buildholder
-	density = 0
-	anchored = 1
+	density = FALSE
+	anchored = TRUE
 	var/client/cl = null
 	var/obj/effect/bmode/builddir/builddir = null
 	var/obj/effect/bmode/buildhelp/buildhelp = null
@@ -144,21 +148,21 @@
 		switch(master.cl.buildmode)
 			if(1)
 				master.cl.buildmode = 2
-				src.icon_state = "buildmode2"
+				icon_state = "buildmode2"
 			if(2)
 				master.cl.buildmode = 3
-				src.icon_state = "buildmode3"
+				icon_state = "buildmode3"
 			if(3)
 				master.cl.buildmode = 4
-				src.icon_state = "buildmode4"
+				icon_state = "buildmode4"
 			if(4)
 				master.cl.buildmode = 1
-				src.icon_state = "buildmode1"
+				icon_state = "buildmode1"
 
 	else if(mods["right"])
 		switch(master.cl.buildmode)
 			if(1)
-				return 1
+				return TRUE
 			if(2)
 				objholder = text2path(input(usr,"Enter typepath:" ,"Typepath","/obj/structure/closet"))
 				if(!ispath(objholder))
@@ -172,9 +176,10 @@
 
 				master.buildmode.varholder = input(usr,"Enter variable name:" ,"Name", "name")
 				if(master.buildmode.varholder in locked && !check_rights(R_DEBUG,0))
-					return 1
+					return TRUE
 				var/thetype = input(usr,"Select variable type:" ,"Type") in list("text","number","mob-reference","obj-reference","turf-reference")
-				if(!thetype) return 1
+				if(!thetype) 
+					return TRUE
 				switch(thetype)
 					if("text")
 						master.buildmode.valueholder = input(usr,"Enter variable value:" ,"Value", "value") as text
@@ -186,7 +191,7 @@
 						master.buildmode.valueholder = input(usr,"Enter variable value:" ,"Value") as obj in object_list
 					if("turf-reference")
 						master.buildmode.valueholder = input(usr,"Enter variable value:" ,"Value") as turf in turfs
-    return 1
+    return TRUE
 
 /proc/build_click(var/mob/user, buildmode, var/list/mods, var/obj/object)
 	var/obj/effect/bmode/buildholder/holder = null
@@ -194,7 +199,9 @@
 		if(H.cl == user.client)
 			holder = H
 			break
-	if(!holder) return
+
+	if(!holder) 
+		return
 
 	switch(buildmode)
 		if(1)
@@ -258,8 +265,7 @@
 				holder.buildmode.objholder = text2path("[object.type]")
 				to_chat(usr, "Selected: [object.type]")
 			else if(mods["right"])
-				if(isobj(object))
-					qdel(object)
+				if(isobj(object)) qdel(object)
 
 		if(3)
 			if(mods["left"]) //I cant believe this shit actually compiles.
